refactor(frontend): extract ydata foundation components
