+++ frontend/ydata-like/index.jsx
@@ -1,327 +1,26 @@
 
+    import { REPORT_DATA } from "./data.js";
+    import { formatBytes, formatPct, formatNum, getAlertColor } from "./helpers.js";
+    import { TypeIcon } from "./components/TypeIcon.jsx";
+    import { HistogramChart } from "./components/HistogramChart.jsx";
+    import { StatRow } from "./components/StatRow.jsx";
+    import { NullityMatrix } from "./components/NullityMatrix.jsx";
+    import { CorrelationMatrixComponent } from "./components/CorrelationMatrixComponent.jsx";
+    import { ScatterPlot } from "./components/ScatterPlot.jsx";
+
     const { useState, useMemo, useEffect } = React;
-    const { 
-      AlertTriangle, 
-      Info, 
-      BarChart2, 
-      Hash, 
-      AlignLeft, 
-      Grid, 
-      Database,
+    const {
+      AlertTriangle,
+      Info,
+      BarChart2,
+      Hash,
+      AlignLeft,
+      Grid,
       ChevronRight,
       ChevronDown,
-      ToggleLeft,
-      AlertCircle,
       PieChart,
       ScatterChart
     } = LucideReact;
-
-    // Securely decode the Base64-encoded payload (Unicode-safe)
-    let REPORT_DATA = { analysis: {}, table: { types: {} }, variables: {}, alerts: [] };
-    try {
-      const ENCODED_REPORT_DATA = "{{REPORT_DATA}}";
-      const bytes = Uint8Array.from(atob(ENCODED_REPORT_DATA), c => c.charCodeAt(0));
-      REPORT_DATA = JSON.parse(new TextDecoder().decode(bytes));
-    } catch (e) {
-      console.error("Error decoding report data:", e);
-    }
-
-    // --- FORMATTING HELPERS ---
-    const formatBytes = (bytes) => {
-      if (!bytes || bytes === 0) return '0 B';
-      const k = 1024;
-      const sizes = ['B', 'KiB', 'MiB', 'GiB', 'TiB'];
-      const i = Math.floor(Math.log(bytes) / Math.log(k));
-      return parseFloat((bytes / Math.pow(k, i)).toFixed(1)) + ' ' + sizes[i];
-    };
-
-    const formatPct = (val) => {
-      if (val === undefined || val === null) return '0.0%';
-      return (val * 100).toFixed(1) + '%';
-    };
-
-    const formatNum = (val) => {
-      if (val === undefined || val === null) return '-';
-      return typeof val === 'number' ? (Number.isInteger(val) ? val.toLocaleString() : val.toFixed(2)) : val;
-    };
-
-    // Helper: Safely parse different matrix JSON structures
-    const parseMatrixData = (data) => {
-      if (!data) return { columns: [], matrix: [] };
-      
-      // 1) Handle { matrix: [ {col1: v1, ...}, ... ] } - new standard
-      if (data.matrix && Array.isArray(data.matrix) && data.matrix.length > 0 && typeof data.matrix[0] === 'object' && !Array.isArray(data.matrix[0])) {
-          const columns = Object.keys(data.matrix[0]);
-          const matrix = data.matrix.map(row => columns.map(col => row[col]));
-          return { columns, matrix };
-      }
-
-      // 2) Handle array of objects format (e.g. from missing heatmap or correlations)
-      if (Array.isArray(data)) {
-        if (data.length === 0) return { columns: [], matrix: [] };
-        const columns = Object.keys(data[0]);
-        const matrix = data.map(row => columns.map(col => row[col]));
-        return { columns, matrix };
-      }
-      
-      // 3) Standard { columns, matrix } or { columns, values }
-      if (data.columns) {
-        if (data.matrix) return { columns: data.columns, matrix: data.matrix };
-        if (data.values) return { columns: data.columns, matrix: data.values };
-      }
-      
-      return { columns: [], matrix: [] };
-    };
-
-    // --- UTILITY COMPONENTS ---
-    const TypeIcon = ({ type }) => {
-      switch (type) {
-        case 'Numeric': return <Hash size={16} className="text-blue-500" />;
-        case 'Categorical': return <AlignLeft size={16} className="text-blue-500" />;
-        case 'Boolean': return <ToggleLeft size={16} className="text-blue-500" />;
-        default: return <Database size={16} className="text-blue-500" />;
-      }
-    };
-
-    const HistogramChart = ({ data, isCategorical = false }) => {
-      if (!data || !data.bins || !data.counts || data.bins.length === 0) return null;
-      
-      const maxCount = Math.max(...data.counts);
-      const total = data.counts.reduce((a, b) => a + b, 0);
-      
-      const chartData = data.bins.map((bin, i) => ({
-        label: bin,
-        count: data.counts[i],
-        percent: ((data.counts[i] / total) * 100).toFixed(1)
-      }));
-
-      if (isCategorical) {
-        return (
-          <div className="flex flex-col gap-3 mt-4 w-full max-w-md">
-            {chartData.map((d, i) => (
-              <div key={i} className="flex flex-col text-sm">
-                <div className="flex justify-between text-gray-600 mb-1">
-                  <span className="truncate pr-4 font-medium">{d.label}</span>
-                  <span className="font-mono text-xs whitespace-nowrap">{formatNum(d.count)} ({d.percent}%)</span>
-                </div>
-                <div className="w-full bg-gray-100 h-5 rounded-sm overflow-hidden flex">
-                  <div 
-                    className="h-full bg-blue-500" 
-                    style={{ width: `${(d.count / maxCount) * 100}%` }}
-                  ></div>
-                </div>
-              </div>
-            ))}
-          </div>
-        );
-      }
-
-      return (
-        <div className="flex items-end h-40 gap-1.5 w-full max-w-md mt-6 mb-10">
-          {chartData.map((d, i) => {
-            const showLabel = i === 0 || i === chartData.length - 1 || i % 5 === 0;
-            return (
-              <div key={i} className="flex-1 flex flex-col justify-end group relative h-full">
-                <div 
-                  className="w-full bg-blue-500 hover:bg-blue-600 transition-colors rounded-t-sm" 
-                  style={{ height: `${Math.max((d.count / maxCount) * 100, 2)}%` }}
-                  title={`${d.label}: ${d.count}`}
-                ></div>
-                {showLabel && (
-                  <div className="text-[11px] text-gray-500 absolute -bottom-6 left-1/2 -translate-x-1/2 whitespace-nowrap">
-                    {String(d.label).replace(/[[\]()]/g, '').split(',')[0]}
-                  </div>
-                )}
-              </div>
-            );
-          })}
-        </div>
-      );
-    };
-
-    const StatRow = ({ label, value, subValue = null, alert = false, compact = false }) => (
-      <div className={`flex justify-between ${compact ? 'py-0' : 'py-1.5'} ${compact ? '' : 'border-b border-gray-100'} last:border-0 text-sm`}>
-        <span className="text-gray-600">{label}</span>
-        <div className="text-right">
-          <span className={`font-medium ${alert ? 'text-red-600' : 'text-gray-800'}`}>{value}</span>
-          {subValue && <span className="text-gray-400 text-xs ml-2">{subValue}</span>}
-        </div>
-      </div>
-    );
-
-    const getAlertColor = (type) => {
-      const colors = {
-        'CONSTANT': 'bg-gray-100 text-gray-800 border-gray-200',
-        'ZEROS': 'bg-orange-100 text-orange-800 border-orange-200',
-        'HIGH_CARDINALITY': 'bg-purple-100 text-purple-800 border-purple-200',
-        'SKEWED': 'bg-yellow-100 text-yellow-800 border-yellow-200',
-        'MISSING': 'bg-red-100 text-red-800 border-red-200',
-        'UNIQUE': 'bg-blue-100 text-blue-800 border-blue-200',
-        'REJECTED': 'bg-slate-200 text-slate-900 border-slate-300'
-      };
-      return colors[type] || 'bg-blue-100 text-blue-800 border-blue-200';
-    };
-
-    // Sub-component for Nullity Matrix Rendering (Missingno style SVG)
-    const NullityMatrix = ({ data, metadata }) => {
-      // Data is now a list of dicts: [ {col1: true, col2: false}, ... ]
-      if (!data || !Array.isArray(data) || data.length === 0) return null;
-
-      const columns = Object.keys(data[0]);
-      const numRows = data.length;
-      const numCols = columns.length;
-      
-      // Convert list of dicts to row-major values array for processing
-      const values = data.map(row => columns.map(c => row[c]));
-
-      const svgWidth = 1000;
-      const svgHeight = 400;
-      const colWidth = svgWidth / numCols;
-      const rowHeight = svgHeight / numRows;
-
-      const rowCompleteness = values.map(row => row.filter(isMissing => !isMissing).length);
-
-      const getBlocks = (cIdx) => {
-          const blocks = [];
-          let inBlock = false;
-          let start = 0;
-          for (let r = 0; r < numRows; r++) {
-              const isPresent = !values[r][cIdx];
-              if (isPresent && !inBlock) {
-                  inBlock = true;
-                  start = r;
-              } else if (!isPresent && inBlock) {
-                  inBlock = false;
-                  blocks.push({ start, end: r - 1 });
-              }
-          }
-          if (inBlock) {
-              blocks.push({ start, end: numRows - 1 });
-          }
-          return blocks;
-      };
-
-      const sparklinePoints = rowCompleteness.map((count, rIdx) => 
-         `${(count / numCols) * 100},${rIdx * rowHeight + rowHeight / 2}`
-      ).join(' ');
-      const sparkPoly = `0,${rowHeight / 2} ${sparklinePoints} 0,${(numRows - 1) * rowHeight + rowHeight / 2}`;
-
-      return (
-        <div className="overflow-x-auto mt-8 pt-4 border-t border-gray-100">
-          <h3 className="text-lg font-medium text-gray-800 mb-2">Matrix</h3>
-
-          {metadata?.truncated && (
-            <div className="mb-4 text-amber-600 text-xs font-medium flex items-center gap-1">
-              <AlertTriangle size={12} /> Matrix truncated to first {metadata.limit} columns out of {metadata.original_count}
-            </div>
-          )}
-
-          <p className="text-sm text-gray-500 mb-6">Visualization of data completeness (first {numRows} rows). Gaps indicate missing data.</p>
-
-          <div className="flex gap-4 items-end" style={{ minWidth: `${Math.max(800, numCols * 25)}px` }}>
-            <div className="flex-1 relative">
-              <div className="flex mb-2">
-                 {columns.map((c, i) => (
-                    <div key={c} style={{ width: `${100/numCols}%` }} className="flex justify-start h-32 relative overflow-visible">
-                       <span className={`absolute bottom-1 left-1/2 origin-bottom-left text-[10px] font-mono text-gray-500 whitespace-nowrap ${numCols > 15 ? '-rotate-90' : '-rotate-45'}`}>{c}</span>
-                    </div>
-                 ))}
-              </div>
-              
-              <svg viewBox={`0 0 ${svgWidth} ${svgHeight}`} preserveAspectRatio="none" className="w-full h-[500px] bg-gray-50 border border-gray-200 shadow-inner rounded-sm">
-                {columns.map((_, cIdx) => (
-                   <g key={cIdx} transform={`translate(${cIdx * colWidth}, 0)`}>
-                      <line x1={colWidth/2} y1="0" x2={colWidth/2} y2={svgHeight} stroke="#e5e7eb" strokeWidth="0.5" strokeDasharray="4 4" />
-                      
-                      {getBlocks(cIdx).map((block, i) => (
-                         <rect
-                            key={i}
-                            x={colWidth * 0.05}
-                            y={block.start * rowHeight}
-                            width={colWidth * 0.9}
-                            height={(block.end - block.start + 1) * rowHeight}
-                            fill="#475569"
-                            className="transition-all hover:fill-blue-500"
-                         />
-                      ))}
-                   </g>
-                ))}
-              </svg>
-            </div>
-
-            <div className="w-20 h-[500px] relative flex flex-col items-center">
-              <div className="text-[10px] text-gray-400 font-mono text-center absolute -top-8 w-full">Data<br/>Completeness</div>
-              <svg viewBox={`0 0 100 ${svgHeight}`} preserveAspectRatio="none" className="w-full h-full border-r border-t border-b border-gray-200 bg-gray-50 rounded-r-sm">
-                <polygon points={sparkPoly} fill="rgba(59, 130, 246, 0.15)" />
-                <polyline points={sparklinePoints} fill="none" stroke="#3b82f6" strokeWidth="1.5" />
-              </svg>
-              <div className="absolute top-2 right-2 text-xs font-mono text-gray-400">{numCols}</div>
-              <div className="absolute bottom-2 right-2 text-xs font-mono text-gray-400">0</div>
-            </div>
-          </div>
-        </div>
-      );
-    }
-
-    // Sub-component for rendering standardized Correlation Matrices
-    const CorrelationMatrixComponent = ({ data }) => {
-      const { columns, matrix } = useMemo(() => parseMatrixData(data), [data]);
-      
-      if (!columns || columns.length === 0 || !matrix || matrix.length === 0) {
-        return <div className="py-12 text-center text-gray-500">No data available for this metric.</div>;
-      }
-
-      return (
-        <div className="overflow-x-auto pb-4">
-          <table className="mx-auto text-xs border-collapse">
-            <thead>
-              <tr>
-                <th className="p-1"></th>
-                {columns.map(c => (
-                  <th key={c} className="p-2 font-mono font-bold text-gray-500 h-28 align-bottom">
-                    <div className={`transform origin-bottom-left ml-6 truncate w-24 ${columns.length > 15 ? '-rotate-90' : '-rotate-45'}`}>
-                      {c}
-                    </div>
-                  </th>
-                ))}
-              </tr>
-            </thead>
-            <tbody>
-              {matrix.map((row, i) => (
-                <tr key={i}>
-                  <td className="p-2 font-mono font-bold text-gray-500 text-right pr-6 whitespace-nowrap">
-                    {columns[i]}
-                  </td>
-                  {row.map((cell, j) => {
-                    const isMissing = cell === null || cell === undefined || Number.isNaN(cell);
-                    const bgOpacity = isMissing ? 0 : Math.min(Math.abs(cell), 1);
-                    const bgColor = isMissing ? 'transparent' : cell >= 0 ? `rgba(59, 130, 246, ${bgOpacity})` : `rgba(239, 68, 68, ${bgOpacity})`;
-                    
-                    return (
-                      <td 
-                        key={j} 
-                        className="w-12 h-12 border border-gray-100 text-center hover:scale-105 transition-transform cursor-crosshair relative group" 
-                        style={{ backgroundColor: bgColor }}
-                      >
-                        {isMissing ? (
-                          <span className="text-gray-400 text-[10px] italic">Null</span>
-                        ) : (
-                          <>
-                            <span className={`font-mono text-[10px] ${bgOpacity > 0.5 ? 'text-white' : 'text-gray-800'}`}>
-                              {cell === 1 ? '1.0' : cell.toFixed(2)}
-                            </span>
-                          </>
-                        )}
-                      </td>
-                    );
-                  })}
-                </tr>
-              ))}
-            </tbody>
-          </table>
-        </div>
-      );
-    };
 
     // --- MAIN SECTIONS ---
     const OverviewSection = ({ data }) => {
@@ -795,82 +494,6 @@
               </table>
             </div>
           </div>
-        </div>
-      );
-    };
-
-    const ScatterPlot = ({ points, xLabel, yLabel }) => {
-      if (!points || points.length === 0) return (
-        <div className="h-96 flex items-center justify-center text-gray-400 italic bg-gray-50 border border-dashed border-gray-200 rounded">
-          No data points available for this pair.
-        </div>
-      );
-
-      const padding = 60;
-      const width = 800;
-      const height = 600;
-
-      const xValues = points.map(p => p.x);
-      const yValues = points.map(p => p.y);
-      const xMin = Math.min(...xValues);
-      const xMax = Math.max(...xValues);
-      const yMin = Math.min(...yValues);
-      const yMax = Math.max(...yValues);
-
-      const xRange = xMax - xMin || 1;
-      const yRange = yMax - yMin || 1;
-
-      const scaleX = (x) => padding + ((x - xMin) / xRange) * (width - 2 * padding);
-      const scaleY = (y) => height - padding - ((y - yMin) / yRange) * (height - 2 * padding);
-
-      return (
-        <div className="flex flex-col items-center w-full">
-          <svg viewBox={`0 0 ${width} ${height}`} className="w-full h-auto bg-white max-h-[700px]">
-            {/* Grid lines */}
-            {[0, 0.25, 0.5, 0.75, 1].map(v => (
-              <React.Fragment key={v}>
-                <line 
-                  x1={padding} y1={height - padding - v * (height - 2 * padding)} 
-                  x2={width - padding} y2={height - padding - v * (height - 2 * padding)} 
-                  stroke="#f1f5f9" strokeWidth="1" 
-                />
-                <line 
-                  x1={padding + v * (width - 2 * padding)} y1={padding} 
-                  x2={padding + v * (width - 2 * padding)} y2={height - padding} 
-                  stroke="#f1f5f9" strokeWidth="1" 
-                />
-              </React.Fragment>
-            ))}
-
-            {/* Axes */}
-            <line x1={padding} y1={height - padding} x2={width - padding} y2={height - padding} stroke="#94a3b8" strokeWidth="2" />
-            <line x1={padding} y1={padding} x2={padding} y2={height - padding} stroke="#94a3b8" strokeWidth="2" />
-
-            {/* Points */}
-            {points.map((p, i) => (
-              <circle 
-                key={i} 
-                cx={scaleX(p.x)} 
-                cy={scaleY(p.y)} 
-                r="3" 
-                fill="#3b82f6" 
-                fillOpacity="0.6"
-                className="hover:fill-blue-800 transition-colors"
-              >
-                <title>{`(${formatNum(p.x)}, ${formatNum(p.y)})`}</title>
-              </circle>
-            ))}
-
-            {/* Axis Labels */}
-            <text x={width / 2} y={height - 10} textAnchor="middle" className="text-xs font-medium fill-gray-500 uppercase tracking-wider">{xLabel}</text>
-            <text x={15} y={height / 2} textAnchor="middle" transform={`rotate(-90, 15, ${height / 2})`} className="text-xs font-medium fill-gray-500 uppercase tracking-wider">{yLabel}</text>
-            
-            {/* Min/Max Labels */}
-            <text x={padding} y={height - padding + 20} textAnchor="middle" className="text-[10px] fill-gray-400 font-mono">{formatNum(xMin)}</text>
-            <text x={width - padding} y={height - padding + 20} textAnchor="middle" className="text-[10px] fill-gray-400 font-mono">{formatNum(xMax)}</text>
-            <text x={padding - 10} y={height - padding} textAnchor="end" className="text-[10px] fill-gray-400 font-mono">{formatNum(yMin)}</text>
-            <text x={padding - 10} y={padding} textAnchor="end" className="text-[10px] fill-gray-400 font-mono">{formatNum(yMax)}</text>
-          </svg>
         </div>
       );
     };
